requirements.txt
flask
python-dotenv
groq
flask-cors
gunicorn
